Revert 11663.
Rev. 11663 causes peers to connect, but then end up backed off due to an AcceptedTimeout. It appears that all but the most basic of communication between nodes is impossible. The node cannot find out its IP address from its peers either. No content is fetchable.

It's late, and I don't have time to look into what exctly the problem is with this patch, so I'll revert it for now to get trunk functional again.


git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@11666 67a373e5-eb02-0410-a15c-ee090a768436