Hopefully fix that
----- Anonymous ----- 2007.03.25 - 18:36:50GMT -----

When accessing a Freenet key that accidentally contains a space at the end, you will get an java.net.MalformedURLException with java.lang.NumberFormatException: For input string: "1 "

This happens e.g. when the key in the clipboard contains a CR-LF at the end and the browser converts this to spaces.

Maybe this could be fixed by removing trailing spaces before processing the key.

git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@12382 67a373e5-eb02-0410-a15c-ee090a768436