Add additional string validity checking functions:
containsNoInvalidCharacters(): checks for lone surrogates and for
FFFE, FFFF, 1FFFE, 1FFFF, etc.

containsNoControlCharacters(): checks for control characters

containsNoInvalidFormatting(): checks for direction control characters
not used in pairs, and interlinear annotation characters not used in
triples.  Neither is really "invalid" formatting, but the first in
particular makes the string dangerous to display in a web page, or any
other sort of user interface.


git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@24693 67a373e5-eb02-0410-a15c-ee090a768436