Added friend feeds. Major modifications to UserAlertManager.java.
It's now possible to recommend files and bookmarks to darknet peers.
The functionality can be accessed both from FProxy and FCP.

FCP support now also exist for Node-to-Node text messages, both
for sending and receiving. When a UserAlert is registerd with the
UserAlertmanager it can send a message to subscribing FCPClients.

Registered alerts can also be accessed from an atom feed file available
at http://127.0.0.1:8888/feed/