Changes to event-handling.
Created an UserEvent-interface. There is now an enum for event-types.
Events are unregistered based on the type. Events can be indefinitely
unregistered. If an event is indefinitely unregistered, not only will
the last registered event be removed, no future events of the same type
that are registered will be displayed.