Statistics and Connections page are pushed
In the statistics page, all but one boxes are pushed in a fixed interval of 10sec. The peers statistics box is pushed when a change occurs.
In the connections page, the peer statistics is pushed when changed, the overview, the activity, the bandwidth and the reject reason boxes are pushed at 10seconds, and the peer lister is pushed at a 10sec interval and when change occurs.