Stop breaking the HTML with arbitrary line breaks.
Line breaks are whitespace in HTML. Additional line breaks that are
introduced in arbitrary places can seriously interefere with the layout of
a page. In this specific case the layout of Freetalk messages can not be
sensibly implemented when something inserts line breaks at will.

This method of keeping the HTML code readable was inspired by Haskell. It
will insert line breaks in the HTML code but as the line breaks are added
within the HTML tags no additional whitespace is rendered.