Try to add more fragments from the same message
Try to add fragments from a message until the packet is full or there is
nothing to add. This helps in cases where the unsent parts of a message
are fragmented (eg. if the middle has been acked), since we won't have
to send two (or more) packets to send the entire message.