Add the WoT version of Freemail as a new official plugin
Since the new version isn't compatible with the old one auto-upgrading
is not a good idea and it was decided to add it as a new plugin. Both
versions can be loaded at the same time, so the only problem with having
both versions visible is that new users might be confused as to which is
the right choice.