Add options for id and to disable probe types.
The options to respond to each probe type default to true, are also
included in test node initialization, and have localization which
includes explanation of what each is and why it can be useful. The
identifier can be manually set or set to the reserved value -1, in which
case it selects a new random value.