Synchronize on counter map; remove stale ones.
The initial block with checking for counters must be synchronized as all
three options should be done as a unit.

Without removing counters when they hit zero the map would slowly
accumulate counters for those from whom it received probe requests; the
majority of which at zero, and likely many for peers no longer connected.
By removing the counter when it hits zero, this is avoided, and as newly
created counters are also zero, this is functionally identical but avoids
a nasty reference leak.