Remove impossible counter check.
The counter is only ever accessed when the counter map is locked, which it
is, so its value cannot change other than from what is run in the current
thread. The mistaken thought process that led to the existence of this
was that if a counter was created but not used, it would not be checked
for removal. However, given that the value cannot change between the two
checks, and being incremented means scheduled decrements, this is not
needed.