Add transport managers
The development history for these can be found in refactor-1 branch.
It is main point of connection between fred plugins and fred for supporting transports.
Synchronization issues haven't been dealt with completely. But performance is not an issue. For now synchrnoized methods are used.
A plugin registers with the registered manager which notifies all the managers present.

A thorough review and good testing is necessary.