Handle transports
This involves creating a new instance of PeerTransport.
Checking noderef if we need that transport.
Storing addresses if we don't have transports.
Locking hasn't been completed.
Performance will be an issue and hence we need separate locks. I haven't thought of the logic yet.
Concurreny will definitely be an issue.