Handle all transports of all peers.
This is a tricky loop. The earlier loop handled all PeerNode maintenance.
Now it handles all transport maintenance for all peers.
The loop does the same thing but for all transports in a PeerNode.

There was some discussion about fairness and loading. But for now this implementation is simple.

Please go through thoroughly. I have to check it once more too.