Remove Freetalk from plugins page.
Even with a warning and only showing the plugin in advanced mode, people
are still installing Freetalk and having problems. Because we do not
want anyone but developers who are working on fixing Freetalk installing
Freetalk, remove the plugin for now.

Once the problems are fixed, the plugin will be added again. Even if
Freetalk were not removed, it would still need a new build of Fred to
distribute the fix.