Fix crash on load of very large bandwidth limits
A very large bandwidth limit was causing an invalid argument to
TokenBucket, and the IllegalArgumentException it throws was not handled
during loading. This showed up when someone entered a long string of 9s
for an incoming bandwidth limit.

It still doesn't proceed to work, as there isn't a default suitable to
all cases, but a controlled exit is better than an uncaught and
unexplained exception.