Rewrite of the SparseBitmap implementation.
The algorithm now make use of the inherent structure of the TreeSet to save time if SparseBitmaps may once be used in a context where more than just a few ranges are contained.

* Complexity of add() was linear in N, is now logarithmic in the total number of ranges N plus linear in the number of merged ranges.
* Complexity of remove() was linear in N, is now logarithmic in N plus linear in the number of removed ranges.
* Complexity of contains() was linear in N, is now logarithmic in N.
* Complexity of notOverlapping() was pseudo-quadratic O((end-start)*N), is now logarithmic in N plus linear in the number of ranges between start->end.

This rewrite also solves an internal flaw where now-overlapping ranges were not merged on add(), and adds Javadoc comments.