Resolved the static method confusion
In freenet.clients.http.ConnectionsToadlet, the static method
drawAddPeersBox is in fred codebase. The previous static style in my
code  was a desperate attempt to fit to the model of fred code base.
Now, I changed drawAddPeerBox to non-static as I see no reason for it to
be static and also the drawNewDarknetPeersAuthBox in ConnectionToadlet
and the corresponding changes