Update translations
Thanks to volunteers from Transifex. Transifex seems to return
untranslated strings even when requested not to. This is not helpful to
have committed, but until they fix it or we write something to avoid it
locally it's easier to just commit it. It will not prevent l10n from
functioning.