Noderefs when user is on 'add a friend'
1) darknetAppServer is final in node
2) After 'add a friend' page is rendered but before user takes any
action, the list should not be modified. Various options using wait,
notify were considered  where we do not update the noderefs list as long
as user stays on 'add a friend' page. But not implemented because I
couldn't find a way to know when the user has refreshed the page/ when
the user clicks on another button i.e. couldn't the finish of this
context.
^ Quick-fix solution to keep a temporary clone of the noderfs list on
this thread is implemented