Ban :visited pseudoclass in CSS filter as it is deemed unsafe.
:visited may change the way a site is displayed based on the browser history of a user, and history information might be disclosed through trusted plugins.
This is described in detail in the comments of ElementInfo.BANNED_PSEUDOCLASS.

We make :visited banned rather than invalid, since the spec prescribes to ignore the entire declaration if a selector is _invalid_, which presents usability issues with disjunctive selectors.
Instead, we ignore only the selector involved if it is deemed valid yet banned.

This also updates CSSParserTest with correct expected results for :visited.