Separate display and SFS/FCP Option representation
Specifically sizes should only be formatted with size units when
displayed to the end user. Conflating display and data transfer/storage
representation, if nothing else, makes FCP more difficult to use. A
bandwidth limit or store size must be parsed for units instead of used
as a number of bytes.