Rename errant localization keys
Property files can look like "key=value", "key = value", "key:value", or
"key value". Fred only recognizes the first, so it can find unescaped
spaces in the middle of a key, whereas conformant parsers find
"key value=more value". Because of this, translators on Transifex are
presented with part of what Fred considerers the key as the start of the
value, and translating it changes the key Fred finds to something the
localization layer isn't queried for.