Split the opennet peers list into long distance peers and short distance peers. LRU operates separately for each. This commit should be fairly mechanistic. See below for more details and attribution. (Doesn't build)
This should fix our link length distribution problems; basically we've had too few short links and too many long links, probably because of path folding on requests for data cached nearby. The result of that appears to have been that routing doesn't work - requests get close via FOAF and then bounce around randomly, and so our closest-approach-distance (on a 5000+ node network) was much too large. We have discussed the poor link length distribution (from probe requests) for some time, but the paper "Measuring Freenet in the Wild: Censorship-resilience under Observation" clarified the situation and gave the impetus for this fix. This is now however exactly their preferred solution, as we are only using 2 "bins"...