FCPPluginClient.sendSynchronous(): Fix memory leak
- The sendSynchronous() adds an object to the synchronousSends map, so
it must also remove it. This is not the job of the other involved thread
which is send(), because it will return after having put the message
into the object in the map, and sendSynchronous() is therefore the last
one to use the object in the map. As the last user, it must remove it.
- While it would theoretically make sense to put this in a separate
try/finally block which starts right after the put() which adds it to
the map, it is not needed:
* remove() will not fail if there is no element with the given key.
* The few calls before the put() should not throw during normal
operation, so it  is not a code path which executes often, and thus does
not have to be optimized.