FCPPluginClient.sendSynchronous(): Fix missing reason for IOException in throws declaration
- It can be thrown if the timeout expires because that might be due to
network problems, which generally are an IOException in standard Java
stuff
- It can be thrown by the backend call to send(), which in turn throws
it if the connection is definitely dead. Thus, we need to tell the user
to consider the connection as dead if this happens.

Notice that I think a timeout should also be considered as a dead
connection as it can indicate that the Internet connection was lost.
This is usually the case with routers which most people use: In my
experience upon connection loss TCP  messages will ping timeout, they
will not be REJECTed. So we should  treat timeout  as a dead connection
to speed up reconnecting.