The two classes FCPPluginMessage: Add/unify JavaDoc which explains the differences between them
- The pluginmanager.* class is the external representation to
server/client plugins, the node.fcp.* is the internal one used by fred.
- The "ATTENTION" is not needed IMHO: The classes provide very different
constructors etc., mixing them up will not compile.
- I have considered renaming the node.fcp one to
"FCPPluginMessageParser" but that would break with the naming of all the
other node.fcp.*Message classes. This is up for discussion though, I
would be willing to rename it. I don't think we should rename the
external class though, FCPPluginMessage fits it perfectly, doesn't it?