PluginRespirator.getPluginClientByID() and backend: Throw IOException instead of PluginNotFoundException
- This function can only be used with the ID of a client, so you have to
have communicated with it already to know about it. Thus, if it does not
exist anymore,  it has disconnected. Network disconnects are signaled
with IOException by convention to match what Java does.
- This has the advantage that you only need a single catch block around
a typical operation of "getPluginClientByID(id).send(...);"
- The backend functions are also changed to use IOException. I have
searched the whole diff of the plugin-fcp-rewrite branch for
PluginNotFoundException. Thus this commit covers all places where the
replacement makes sense, and so it removes the FIXME which demanded this
change at FCPPluginClient. Notice that the send-functions of
FCPPluginClient do not have to be changed, they use IOException already.
- Also, the exception is amended with an error message and the ID of the
client which wasn't found.