Rename class FCPPluginReply to FCPPluginServerMessage
- The idea behind this is the problem that FCPPluginReply is not just a
reply anymore: The old plugin FCP code did only allow the server to send
messages to a client as an immediate answer to a message from the
client. With the changes of this branch, the server is allows to store
an ID of the client and contact the client on his own. So the "reply"
needs to go away.
- Notice that the class "FCPPluginClientMessage" which FCPPluginReply's
new JavaDoc references doesn't exist yet. It shall be the new name of
class freenet.node.fcp.FCPPluginMessage, which will be renamed to that
name in the following commit. I decided to use that name in the JavaDoc
already because otherwise the symmetry of names it talks about doesn't
make sense.
- The FIXME at FCPPluginReply asked to add a FIXME  about adding a
duplicate on-network name for FCPPluginServerMessage/FCPPluginReply to
its static String NAME. The FIXME was added, but to the class-level
JavaDoc. It is better to explain the mismatch between NAME and class
name right at the class level because it is important to know.