Rename class freenet.node.fcp.FCPPluginMessage to FCPPluginClientMessage
- This is to remove the name clash with
FredPluginFCPMessageHandler.FCPPluginMessage
- This class is renamed instead of the other one because it has a more
limited purpose: It is only used to represent client-to-server messages.
The other one is a generic purpose representation for both
client-to-server and server-to-client.
- The FIXME to change
anything which uses "FredPluginFCPMessageHandler.FCPPluginMessage"
instead of "FCPPluginMessage" due to the nameconflict is postponed to
the following commit since it will be a large amount of changes.
Meanwhile, the FIXME is moved to the FredPluginFCPMessageHandler.* class
because its better placed there.