FCPPluginClientMessage: Remove FIXME of adding a "IsReply" / "IsAnswer" field to the on-network format
I've thought for quite some time about this. Reasons against it:
- It would only add additional nesting of the already triple-nested
logic  which checks  whether there is a Success field and  if there is
checks  whether there is an errorCode field and if there is one of that
checks  whether there is an errorMEssage feild.
- Make "FCPPluginReply" FCP messages (represented by class
FCPPluginServerMessage) confusing: They would possibly contain a
IsReply=false field becaue servers can send messages to clients without
prior client action if an event happens at the server. This is
especially bad because there is no easy way to rename FCP messages
which would allow us to rename FCPPluginReply: The only way of adding a
secondary name for the transition period seems to be to introduce a
second message type with the new name and send a duplicate of the whole
message. That sucks.
- It was suggested by the FIXME to add the IsReply field because doing
the logic implicitely with checking whether there is a "Success" field
is suspected to cause server/client authors to leave it out because they
think its not needed. But the requirement of having the IsReply field
would be communicated to the authors the same way as the Success field:
Via the documentation of FCP. If they chose to ignore it for the Success
field, they will as well for the IsReply field.