FCPPluginClient.sendSynchronous(): Increase default timeout assert() limit
- I think it is enough that we have a 19 line JavaDoc at
sendSynchronous() which tells people to use it in a sane manner, we
shouldn't need excessive assert() policing.
- The current implemention of WOT event-notifications creates a maximum
of 3 sendSendSynchronous() threads when starting an event-subscription
to send huge amount of bulk data. It will use Buckets for that. While
this should transfer quickly for local connections, it might be slow
with networked FCP, and testing VMs. So the thread amount will be low,
but their lifetime can be high. I agree that we need to police
sendSynchronous() calls to prevent excessive thread growth. But stuff
which programmers can program into plugins have full "root" access to
anything in Freenet anyway so you can always shoot yourself in the feet.