FCPPluginClient: Add constructor for unit tests
- This will allow us to adapt the existing WOT unit tests which use the
old FCP API to also work with the new FCPPluginClient.
- The backend constructor which is being used is the same as in
constructForIntraNodeFCP(). This seems the right thing to do because we
are running in a single VM, so it sort of *is* intra-node FCP.
- I am NOT sure whether PooledExecutor is the right thing here. I merely
searched the test directory of fred for Executor and observed that it is
used in many places. I'll just try it in WOT and see whether it blows
up.