NodeStarter.TestNodeParameters: Make it clear how the node directory is specified
- Instead of having a "String testName" which does not suggest that it
actually is the directory name, use a "File basedirectory".
- Explain that people should ensure that this does not exist already
when doing tests so they don't re-use the existing data. Notice the
claim "the behavior of {@link File#mkdir()} is undefined for existing
directories." stems from the fact that the JavaDoc doesn't state what
happens if the directory exists already. It does fail for me then, but
unless the JavaDoc states that, it should not be taken for granted.
- Default the directory name to use a random UUID as part of the name to
ensure that collisions won't happen if people don't specify a directory.
- The existing code paths are not touched, they will stay with a String
testName.
- This will be needed for adapting WOT unit tests to use
plugin-fcp-rewrite: To ensure that the WOT database does not persist
across test restarts, the best check is to check whether the fred
directory does not exist. This is only possilbe if WOT knows what it is
*before* it has even created the test-Node: The node will create the
directory during startup, so querying it from it after startup is
useless as it exists then already.