NodeStarter.TestNodeParameters: Fix wrong JavaDoc which I had added
The said createTestNode() does the following:
  if((!portDir.mkdir()) && ((!portDir.exists()) ||
      (!portDir.isDirectory()))) {
	fail(); ...

This clearly does not fail if the directory exists already. Sorry.
So the JavaDoc is corrected to state that it will not fail if the dir
exists.
Also, it is adapted to point the user to the name of the JUnit class for
automatic directory management.