NodeStarter.globalTestInit(): Make parameters more self-explanatory
- The previous "testName" was only used for creating the test directory
from it, so it is better to consume a "File baseDirectory" instead. This
makes it coherent with the similar change in TestNodeParameter.
- Also, add JavaDoc to the baseDirectory of TestNodeParameters to
explain the interaction with globalTestInit(): globalTestInit() creates
the parent directory baseDirectory, which then is also specified in
TestNodeParameters, and the child nodes will put their data into a
subdirectory in it with the name being their port number.