Node: Fix ClassCastException when using DummyRandomSource instead of Yarrow
- ClassCastException would happen if the recent modifications of
NodeStarter.createTestNode() were used to run a Node which uses a
DummyRandomSource as RNG instead of Yarrow: During shutdown, the node
would try to cast the RandomSource to Yarrow so it can call write_seed()
upon it
- The fix is to:
* introduce an interface PersistentRandomSource
* make Yarrow implement it, but not DummyRandomSource.
* Check whether the RandomSource is a PersistentRandomSource during
shutdown, and only cast to it if it is. Otherwise, do not try to write a
seed file as a seed file from a DummyRandomSource is worthless.