PluginRespirator.connectToOtherPlugin(): JavaDoc how to disconnect
- This was previously only explained in the "Internals" section of
FCPPluginClient, so users were likely to not see it. It is a better
place to put it at the connect*() function because if you connect you
always need to know hwo to disconnect.
- sendSynchronous() is suggested for pinging for the following reason:
While it will block the ping loop thread and thus require a thread to be
alive, this is what a regular send() implementation would probably also
end up like: With a regular send(), you should send() the ping, then do
Thread.sleep(), and check afterwards whether the pong has arrived. So
the thread would be blocked. But in addition to that, you would need a
variable in your class which is also accessible to the
FredPluginFCPMessageHandler where it can store that a ping was received.
Using sendSynchronous() eliminates the variable, thus reduces complexity
of the class, and has the same impact upon thread usage. Thus, it is
recommended.