FCPPluginClient: Implement logging of all messages
This is implement at the highest level of code, i.e. it will log
FCPPluginMessage objects. This means that if the messages are serialized
into network data, that data will NOT be logged.
To ensure that this works properly, we should  also implement
unit tests for FCPPluginServerMessage / FCPPluginClientMessage which
ensure that the serialization of messages to network data does not break
anything. Will do so in the following commits.