FCPPluginClientMessage: Prevent NPEs in server code
- Hand out a mock SimpleFieldSet if the client did not specify one

- This does change the legacy codepath of the PluginTalker stuff, which
this branch aims to NOT modify. But IMHO this modification should not
have any bad effects:
* Null SimpleFieldSet were not normal mode of operation there because
the SimpleFieldSet was the only payload of the message. In other words:
The new plugin-fcp-rewrite code which replaces PluginTalker also has
the success / errorCode / errorMessage payload, so there is
justification for having a null SimpleFieldSet.
* Changing something from null to non-null is likely to not break much,
it will rather prevent stuff from breaking.

- Fixes the failure of FCPPluginMessageEncodeDecodeTest:#

[junit] java.lang.NullPointerException
[junit]     at freenet.node.fcp.FCPPluginMessageEncodeDecodeTest
		.testEncodeDecode(FCPPluginMessageEncodeDecodeTest.java:136)
[junit]     at freenet.node.fcp.FCPPluginMessageEncodeDecodeTest
		.testEncodeDecode(FCPPluginMessageEncodeDecodeTest.java:88)