Location: use Location objects instead of doubles
This commit replaces most double-typed locations with Location objects. Most
important exception is network-level messages. String representation of the new
Location objects remains unchanged (i.e. textual double representation).

The general idea is that only very low-level parts of the code have to have
knowledge on the type (and peculiarities) of the actual location encoding, most
parts can do with some simple operations on locations (e.g. distance, isValid).
Those operations are now encapsulated in the Location objects, instead of
providing them as static operations on doubles. The old static Location methods
are converted to work on Location objects, and are deprecated for future removal
(this causes a lot of warnings at the moment).

This changes mainly serve to prevent errors in duplication of location logic.
Apart from this, it may also serve as preliminary work for switching the
internal representation of locations from doubles to longs, if that may be
considered necessary at some point.

To do:
* documentation of introduced methods
* removal of deprecated static methods