Node config defaults: fix minimum store size bug
The storeSize configuration option accepted values down to 10M, but the
node would fail to start when the store size was less than 32MiB.

This commit consolidates the knowledge on minimum store sizes to
constants in Node, raising the minimum datastore size to 32MiB in the
configuration options. This commit also clarifies on error messages for
invalid store sizes, and adds unit tests for sanity of default store
sizes.