Fix exception when parsing “/*/*”
When a combination of / and * is encountered, the parser switches to
STATECOMMENT; then it continues to evaluate the current character. This
leads to the * that started the comment also finishing the comment. On
the next occurence of / and * another comment is started and the slash
that is supposed to be in the current buffer is removed; as the buffer
is empty at this time, this subsequently fails.

The fix makes sure that when switching to or from STATECOMMENT the
current character is not evaluated again, mainly by clearing out the
“prevc” variable which contains the previous character.