Introduce lightweight exceptions without stack trace
Exceptions are expensive because filling in their stack traces is
expensive. If we use exceptions for control flow, we can avoid
filling stack traces that will never be used. This commit introduces
LightweightExceptions that do not provide a stack trace by default,
and rewrites existing exceptions that (conditionally) did not provide
stack traces into subclasses of LightweightException.