WrapperConfig: always set reload_configuration=TRUE
WrapperConfig unintentionaly failed to write
wrapper.restart.reload_configuration=TRUE to wrapper.conf on property
write (but only when it was set), causing the configuration to not be
reloaded on automated restart of the wrapper.

This caused stale values to be used and may have resulted in loading
the wrong (previous) version on auto-update, in turn causing an
infinite update loop.