WIP: progress toward applied update notification
This is an incomplete implementation. There are a few ways to go from
here:

* get rid of the new configuration options and just show the alert if
  lastVersion is different. This has the disadvantage of only showing
  on first startup after an upgrade.
* get rid of the "don't show these" button localization because
  implementing it is likely to be painful. It'd be great from a
  usability perspective though. Maybe halfway would be to link to the
  config page anchor of the relevant setting like other alerts do?

This also adds a marker for a likely place to add parsing the manifest
of the update jar for compatibility and other information.

The links can be added to this alert with the same function used in the
pending upgrade notification and the statistics page:
NodeUpdaterManager.addChangelogLinks().