FCPPluginClient: Don't fake missing replies, log instead.
The WOT FCPInterface uses the case of not sending a reply to prevent the
client form hammering it with retries during shutdown. This is a valid
usecase.
Also, impersonating the message handler by faking replies would be
impolite and would cause developers to run into headaches when trying to
figure out why they receive those empty success replies.