Default bookmarks: fix operhiem1's blog not appearing
This was added in 20f1363, but we did not realize that the default
bookmark file has an explicit length, and the additional link was beyond
it and so is not displayed. The more thorough fix is to determine the
length from the input instead of specifying it explicitly, but this
change is sufficient for the time being.