Update dependencies in ant and documentation
JUnit4 and Hamcrest are required, as is Bouncy Castle 1.51. The Bouncy
Castle filename used now includes the version for clarity. Libraries are
searched for in the system-wide /usr/share/java by default, so
installing dependencies from packages is sufficient.

freenet.jar no longer defines a Main-Class or a Class-Path because a
node cannot be started with -jar:

1) It prevents specifying additional libraries on the classpath, which
   Freenet requires. From the java man page:

When you use this option, the JAR file is the source of all user
classes, and other user class path settings are ignored.

2) If the dependencies were not a problem it would imply that default
   JVM settings on such things as memory limits are sufficient, and they
   are likely to be too low.