FCPPluginClient: Document how sendSynchronous() shutdown works
I don't think its worthwhile to implement a table which keeps a list of
the threads. I'm rather delegating the job of keeping track of the
threads to the plugin authors by having this commit add JavaDoc which
requests plugin authors to
- keep track of threads running sendSynchronous() on their own
- calling Thread.interrupt() upon them at shutdown of their plugin

(Notice: Fred also won't interrupt() all threads of plugins during
shutdown, I've searched the code. So its really the job of plugin
authors to deal with this)