PluginRespirator.connectToOtherPlugin(): JavaDoc
Remove recommendation of pinging by FCPPluginClient.sendSynchronous().
sendSynchronous() is a rather heavy function as it keeps the thread
alive.
Pinging can be done without a live thread by having a timer which wakes
up the keepalive loop, and storing the "last time a ping was received"
date in a member variable. No need for a thread there.
Also, I guess explaining how to implement pinging is beyond the scope of
the JavaDoc of this function anyway - its basic networking knowhow.