ClientSideFCPMessageHandler: JavaDoc
Do not claim that messages which servers send on their own have a unique
identifier:
This is not required in the ServerSideMessageHandler JavaDoc nor
anywhere else.
It pretty much depends on the implementation of the server whether it
sends its own messages as replies or as non-reply messages with an
unique identifier. It wouldn't make sense to put restrictions upon that
as whats needed depends on the specific protocol which is being
implemented.