Post-merge fixes of merging purge-db4o
Move class FCPPluginClientMessage from package freenet.node.fcp to
freenet.clients.fcp as the related classes were moved like that as well
in purge-db4o.

Notice: This intentionally does NOT add the following imports from the
old package:
	import freenet.node.fcp.DataCarryingMessage;
	import freenet.node.fcp.ProtocolErrorMessage;
This is because it seems like they were copy-pasted to the new package
as part of toad's purge-db4o work, and the ones in the old package are
obsolete, simplified versions which merely are still existing so the
node can import old db4o databases.

This is part of a series of commits to fix compiler issues as a result
of the merge 85be221b19882802806ca12614e96ad156f466d6.
The series is not complete yet, the last commit will tell its the last.