Post-merge fixes of merging purge-db4o
Cherry-pick changes which were lost due to the fact that Git did not
cleanly handle the rename-rename conflict of:
- freenet.node.fcp.FCPPluginMessage, which was renamed to
freenet.clients.fcp.FCPPluginMessage in purge-db4o, and now is
freenet.clients.fcp.FCPPluginClientMessage
- freenet.node.fcp.FCPPluginReply, which was renamed to
freenet.clients.fcp.FCPPluginReply in purge-db4o, and now is
freenet.clients.fcp.FCPPluginServerMessage

I manually viewed the Git history of the files in purge-db4o, and
determined which commits were done onto them in between. It merely was
commit 19fa0fac48d339d665f88b50c30c44e73d334aaa. Thus, this commit
manually applies the changes of that one.