Post-merge fixes of merging purge-db4o
Cherry-pick changes from plugin-fcp-rewrite which were lost due to the
fact that class FCPMessage was copy-pasted to a duplicate one by
purge-db4o:
There is freenet.clients.fcp.FCPMessage and freenet.node.fcp.FCPMessage
now.
During merging, Git wanted to apply the changes to the freenet.node one,
but that did not work because it doesn't contain the affected code
anymore.
The proper solution is to apply the changes to the freenet.clients one,
which this commit does.

I manually viewed the Git history of the file in plugin-fcp-rewrite, and
determined which commits were done onto it. It merely was commit
667f74ed462155e50dcebf16314341b7577552a8. Thus, this commit manually
applies that one.